reorganize button components
