+++ frontend/apps/marketing/src/components/contentful/button/Button.tsx
@@ -1,67 +1,53 @@
+import OpenInNew from '@mui/icons-material/OpenInNew';
+import MuiButton from '@mui/material/Button';
+import classNames from 'classnames';
 import {EntryFields} from 'contentful';
-import React, {useMemo} from 'react';
-
-import {
-  ButtonColor,
-  ButtonType,
-  LinkButton,
-} from '@code-dot-org/component-library/button';
-
-import {
-  externalLinkIconProps,
-  fontAwesomeV6BrandIconsMap,
-} from '@/components/common/constants';
 
 type ButtonProps = {
   /** Button text */
   text?: string;
-  /** Button color */
-  color: Extract<ButtonColor, 'purple' | 'black' | 'white'>;
   /** Button type (semantic) */
-  type: Extract<ButtonType, 'primary' | 'secondary'>;
+  type: 'emphasized' | 'primary' | 'secondary';
+  /** Button size */
+  size: 'small' | 'medium' | 'large';
   /** Button link href */
   href?: string;
   /** Whether Link is for internal code.org pages, or external web page. (external links are opened in new tab) */
   isLinkExternal?: boolean;
-  /** Button left icon name */
-  iconLeftName?: string;
   /** Aria label for the button */
   ariaLabel?: EntryFields.Text;
+  /** Custom classname */
+  className?: string;
 };
 
 const Button: React.FunctionComponent<ButtonProps> = ({
   text,
-  color,
   type,
+  size,
   href,
   isLinkExternal = false,
   ariaLabel,
-  iconLeftName,
+  className,
 }) => {
-  const isLeftIconBrand = useMemo(
-    () => !!iconLeftName && fontAwesomeV6BrandIconsMap.has(iconLeftName),
-    [iconLeftName],
-  );
   return (
-    <LinkButton
-      text={text}
-      size="m"
-      href={href}
-      target={isLinkExternal ? '_blank' : '_self'}
-      type={type}
-      color={color}
-      aria-label={ariaLabel}
-      iconLeft={
-        iconLeftName
-          ? {
-              iconName: iconLeftName,
-              iconStyle: 'solid',
-              iconFamily: isLeftIconBrand ? 'brands' : undefined,
-            }
-          : undefined
-      }
-      iconRight={isLinkExternal ? externalLinkIconProps : undefined}
-    />
+    href && (
+      <MuiButton
+        className={classNames(`button--color-${type}`, className)}
+        variant={
+          type === 'emphasized' || type === 'primary' ? 'contained' : 'outlined'
+        }
+        size={size}
+        href={href}
+        target={isLinkExternal ? '_blank' : undefined}
+        rel={isLinkExternal ? 'noopener noreferrer' : undefined}
+        aria-label={ariaLabel}
+        endIcon={isLinkExternal ? <OpenInNew /> : undefined}
+        disableElevation
+        disableRipple
+      >
+        {text}
+      </MuiButton>
+    )
   );
 };
