+++ frontend/apps/marketing/src/components/contentful/button/index.ts
@@ -1,3 +1,3 @@
 // Export the Component Definition for use in Contentful Studio
-export {ButtonContentfulComponentDefinition} from './ButtonContentfulDefinition';
+export {ButtonCSforAllContentfulComponentDefinition} from './ButtonCSforAllContentfulDefinition';
 export {default} from './Button';
